fix Incorrect display of Number picker type

+++ app/src/main/java/night/app/data/entities/Sleep.java
@@ -10,10 +10,10 @@ public class Sleep {
     public Integer day;
 
     @NonNull
-    public Integer startTime;
+    public Long startTime;
 
     @NonNull
-    public Integer endTime;
+    public Long endTime;
 
     public Integer enableMission;
